Knife leveling system fixed

+++ src/main/java/net/LiorNadav/rpgmod/weapon_leveling_system/warrior/knife/PlayerKnife.java
@@ -57,7 +57,7 @@ public class PlayerKnife{
         return knifeExperience;
     }
 
-    public float getKnifeLevel(){
+    public int getKnifeLevel(){
         return knifeLevel;
     }
